requirements.txt
PySide6>=6.5.0
PySide6-WebEngine>=6.5.0
Pillow>=9.0.0
rawpy>=0.17.0
exifread>=3.0.0
geopy>=2.3.0
requests>=2.28.0 